Fix incorrect interpretation of transforms.
The examples were tested more closely and changes were made.
The transformations aren't compounded onto one another.
And it's not the keyboard that is shifting, it's a target pointer that
moves position for every instruction.  And each key position processes
all instructions before the next key pointer is calculated.

Previously, the keys themselves were shifting positions, and the entire
keyboard would be transformed before the next transformatoin instruction
was considered.  This is a much more complicated encoding but is not
correct as per the examples.